All the DBI errors have now a DBI specific error code. WIP - advanced in DBI module side management of blob fields (almost done).
git-svn-id: svn://falconpl.org/falcon/modules/dbi/trunk@83 6551dcfb-e141-0410-b529-ba959add1974

+++ sqlite3/sqlite3_srv.cpp
@@ -205,9 +205,9 @@ dbi_status DBIRecordsetSQLite3::asDate( const int columnIndex, TimeStamp &value 
    tv.subString( 5, 7 ).parseInt( month );
    tv.subString( 8, 10 ).parseInt( day );
 
-   value.m_year = year;
-   value.m_month = month;
-   value.m_day = day;
+   value.m_year = (int16) year;
+   value.m_month = (int16) month;
+   value.m_day = (int16) day;
    value.m_hour = 0;
    value.m_minute = 0;
    value.m_second = 0;
@@ -241,9 +241,9 @@ dbi_status DBIRecordsetSQLite3::asTime( const int columnIndex, TimeStamp &value 
    value.m_year = 0;
    value.m_month = 0;
    value.m_day = 0;
-   value.m_hour = hour;
-   value.m_minute = minute;
-   value.m_second = second;
+   value.m_hour = (int16) hour;
+   value.m_minute = (int16) minute;
+   value.m_second = (int16) second;
    value.m_msec = 0;
 
    return dbi_ok;
@@ -274,12 +274,12 @@ dbi_status DBIRecordsetSQLite3::asDateTime( const int columnIndex, TimeStamp &va
    tv.subString( 14, 16 ).parseInt( minute );
    tv.subString( 17, 19 ).parseInt( second );
 
-   value.m_year = year;
-   value.m_month = month;
-   value.m_day = day;
-   value.m_hour = hour;
-   value.m_minute = minute;
-   value.m_second = second;
+   value.m_year = (int16) year;
+   value.m_month = (int16) month;
+   value.m_day = (int16) day;
+   value.m_hour = (int16) hour;
+   value.m_minute = (int16) minute;
+   value.m_second = (int16) second;
    value.m_msec = 0;
 
    return dbi_ok;
